[Chore/#165] TabBar AppTab Model -> TabBarVC 상단으로 이동
